cleanup: move storage functions to storage module.
In the context of cleanup instance make some refactoring (there is no new logic implemented):
 - move the storage functions to their own to standalone module "storage"
 - move the storage functions unit-tests to "storage" module"
 - add a unittest for "GetPathFromURL"
This is a preparation work for cleanup devices module.
FIXES: https://issues.redhat.com/browse/THEEDGE-3489
